+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 
+### Changes
+* Updated Kurtosis Kubernetes `label_key_consts` and tests.
+
 # 0.20.1
 ### Features
 * Added `KubernetesKurtosisBackend.GetEnclaves` functionality to kubernetes backend